requirements.txt
discord
rich
feedparser
beautifulsoup4
requests
pytz
python-dotenv